+++ src/UniqueSluggifier.php
@@ -10,7 +10,7 @@
  * @package caseyamcl/toc
  * @author Casey McLaughlin <caseyamcl@gmail.com>
  *
- * For the full copyright and license information, please view the LICENSE
+ * For the full copyright and license information, please view the LICENSE.md
  * file that was distributed with this source code.
  *
  * ------------------------------------------------------------------